+++ code/backend/controllers/admin/org.go
@@ -1,21 +1,22 @@
 package admin
 
 import (
-	"github.com/temphia/temphia/code/backend/libx/easyerr"
 	"github.com/temphia/temphia/code/backend/xtypes/models/claim"
 	"github.com/temphia/temphia/code/backend/xtypes/models/entities"
+	"github.com/temphia/temphia/code/backend/xtypes/scopes"
 )
 
 func (c *Controller) GetTenant(uclaim *claim.Session) (*entities.Tenant, error) {
-	if !uclaim.IsSuperAdmin() {
-		return nil, easyerr.NotAuthorized()
+	if !c.HasScope(uclaim, "tenant") {
+		return nil, scopes.ErrNoAdminTenantScope
 	}
+
 	return c.coredb.GetTenant(uclaim.TenantId)
 }
 
 func (c *Controller) UpdateTenant(uclaim *claim.Session, data map[string]any) error {
-	if !uclaim.IsSuperAdmin() {
-		return easyerr.NotAuthorized()
+	if !c.HasScope(uclaim, "tenant") {
+		return scopes.ErrNoAdminTenantScope
 	}
 
 	delete(data, "slug")
@@ -26,22 +27,42 @@ func (c *Controller) UpdateTenant(uclaim *claim.Session, data map[string]any) er
 // domain
 
 func (c *Controller) AddDomain(uclaim *claim.Session, domain *entities.TenantDomain) error {
+	if !c.HasScope(uclaim, "tenant") {
+		return scopes.ErrNoAdminTenantScope
+	}
+
 	domain.TenantId = uclaim.TenantId
 	return c.coredb.AddDomain(domain)
 }
 
 func (c *Controller) UpdateDomain(uclaim *claim.Session, id int64, data map[string]any) error {
+	if !c.HasScope(uclaim, "tenant") {
+		return scopes.ErrNoAdminTenantScope
+	}
+
 	return c.coredb.UpdateDomain(uclaim.TenantId, id, data)
 }
 
 func (c *Controller) GetDomain(uclaim *claim.Session, id int64) (*entities.TenantDomain, error) {
+	if !c.HasScope(uclaim, "tenant") {
+		return nil, scopes.ErrNoAdminTenantScope
+	}
+
 	return c.coredb.GetDomain(uclaim.TenantId, id)
 }
 
 func (c *Controller) RemoveDomain(uclaim *claim.Session, id int64) error {
+	if !c.HasScope(uclaim, "tenant") {
+		return scopes.ErrNoAdminTenantScope
+	}
+
 	return c.coredb.RemoveDomain(uclaim.TenantId, id)
 }
 
 func (c *Controller) ListDomain(uclaim *claim.Session) ([]*entities.TenantDomain, error) {
+	if !c.HasScope(uclaim, "tenant") {
+		return nil, scopes.ErrNoAdminTenantScope
+	}
+
 	return c.coredb.ListDomain(uclaim.TenantId)
 }
